requirements.txt
aiogram==2.12.1
aiohttp==3.7.4
pydantic==1.8.1
python-dotenv==0.16.0
